enhanced txn tests
